+++ tests/test_files.py
@@ -2,10 +2,16 @@ import os
 import tempfile
 import unittest
 import shutil
-from satella.files import read_re_sub_and_write, find_files
+from satella.files import read_re_sub_and_write, find_files, split
 
 
 class TestFiles(unittest.TestCase):
+    def test_split(self):
+        self.assertEqual(split('c:/windows/system32/system.exe'), ['c:/', 'windows', 'system32',
+                                                                   'system32.exe'])
+        self.assertEqual(split('~/user/something/./else'), ['~', 'user', 'something', '.',
+                                                            'else'])
+
     def setUp(self) -> None:
         self.filename = tempfile.mktemp()